reverting changes to serial port types
